# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test task API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/company": {
            "get": {
                "tags": [
                    "Company"
                ],
                "description": "Get list of all companies",
                "operationId": "3e0bffb951b3f31f83aadc004638122b",
                "responses": {
                    "200": {
                        "description": "List of companies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CompanyDTO"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "http": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Company"
                ],
                "description": "Create a company",
                "operationId": "faf1545d2af89079ca55c4f9ed6ea302",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCompanyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Company created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Cannot find building by buildingId"
                    },
                    "409": {
                        "description": "Id already taken"
                    }
                },
                "security": [
                    {
                        "http": []
                    }
                ]
            }
        },
...
        "/api/company/search": {
            "post": {
                "tags": [
                    "Company"
                ],
                "description": "Search companies by name",
                "operationId": "bec748b9b6ae5814dc3823ee46fbd6ff",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SearchCompanyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List of companies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CompanyDTO"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "http": []
                    }
                ]
            }
        },
        "/api/company/{id}": {
            "get": {
                "tags": [
                    "Company"
                ],
                "description": "Get info about specific company",
                "operationId": "29b0f8f11c5e74b0c7a9f94f223c0abc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cannot find company by id"
                    }
                },
                "security": [
                    {
                        "http": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Company"
                ],
                "description": "Update a company",
                "operationId": "dba5691bd7818c368accfd785048ea90",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCompanyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Company updated"
                    },
                    "404": {
                        "description": "Cannot find company by id"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "http": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Company"
                ],
                "description": "Delete a company",
                "operationId": "a073402836212623826fbe1e5a591454",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company deleted"
                    },
                    "404": {
                        "description": "Cannot find company by id"
                    }
                },
                "security": [
                    {
                        "http": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "CompanyDTO": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "buildingId": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "CreateCompanyRequest": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "buildingId": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "SearchCompanyRequest": {
                "properties": {
                    "search": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateCompanyRequest": {
                "properties": {
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "http": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}